+++ lockbox-ios/Action/SentryAction.swift
@@ -6,7 +6,6 @@ import Foundation
 
 struct SentryAction: Action {
     let title: String
-    let error: Error
-    let function: String
-    let line: String
+    let error: Error?
+    let line: String?
 }